+++ R/hyper_tbl_cube.R
@@ -5,6 +5,10 @@
 #' into the tbl cube format defined in the [dplyr][cubelyr::tbl_cube] 
 #' package. 
 #' 
+#' The size of an extraction is checked and if *quite large* there is an a user-controlled
+#' prompt to proceed or cancel. This can be disabled with `options(tidync.large.data.check = FALSE)` 
+#' - please see [hyper_array()] for more details. 
+#' 
 #' The tbl cube is a very general and arbitrarily-sized array that 
 #' can be used with tidyverse functionality. Dimension coordinates are
 #' stored with the tbl cube, derived from the grid 
@@ -12,6 +16,7 @@
 #'
 #' @param x tidync object
 #' @param ... arguments for [hyper_filter()]
+#' @param force ignore caveats about large extraction and just do it
 #' @seealso [hyper_array()] and [hyper_tibble()] which are also delay-breaking 
 #' functions that cause data to be read 
 #' @return tbl_cube
@@ -37,13 +42,13 @@
 #' }
 #' @return `dplyr::tbl_cube`
 #' @export
-hyper_tbl_cube <- function(x, ...) {
+hyper_tbl_cube <- function(x, ..., force = FALSE) {
   UseMethod("hyper_tbl_cube")
 }
 #' @name hyper_tbl_cube
 #' @export
 #' @importFrom stats setNames
-hyper_tbl_cube.tidync <- function(x, ...) {
+hyper_tbl_cube.tidync <- function(x, ..., force = FALSE) {
   active_names <- tibble::tibble(dim = as.integer(
     gsub("^D", "", unlist(strsplit(active(x), ",")))))
   dim_names <- active_names %>% 
@@ -57,14 +62,14 @@ hyper_tbl_cube.tidync <- function(x, ...) {
       dplyr::pull(inm)
   }
   ldims <- lapply(dim_names, lfun)
-  structure(list(mets = hyper_array(x, ...), 
+  structure(list(mets = hyper_array(x, ..., force = force), 
                  dims = setNames(ldims, dim_names)), 
             class = "tbl_cube")
 }
 #' @name hyper_tbl_cube
 #' @export
-hyper_tbl_cube.character <- function(x, ...) {
-  tidync(x) %>% hyper_filter(...) %>% hyper_tbl_cube()
+hyper_tbl_cube.character <- function(x, ..., force = FALSE) {
+  tidync(x) %>% hyper_filter(...) %>% hyper_tbl_cube(force = force)
 }