+++ R/hyper_tibble.R
@@ -1,11 +1,15 @@
 #' Extract NetCDF data as an expanded table.
 #'
 #' Extract the raw array data as an expanded data frame. This can be the entire
-#' variable/s or after dimension-slicing using `hyper_filter` expressions with
+#' variable/s or after dimension-slicing using [hyper_filter()] expressions with
 #' dimension values expanded appropriately for each element in the arrays (one
 #' row per element).
 #'
-#' The function `hyper_array` will act on an existing tidync object or a source
+#' The size of an extraction is checked and if *quite large* there is an a user-controlled
+#' prompt to proceed or cancel. This can be disabled with `options(tidync.large.data.check = FALSE)` 
+#' - please see [hyper_array()] for more details. 
+#' 
+#' The function [hyper_tibble()] will act on an existing tidync object or a source
 #' string.
 #'
 #' By default all variables in the active grid are returned, use `select_var` to
@@ -15,6 +19,7 @@
 #' @param ... arguments to `hyper_filter``
 #' @param na.rm if `TRUE` these rows are not included in the output when all
 #'   variables are `NA`
+#' @param force ignore caveats about large extraction and just do it
 #' @return a `tbl_df`
 #' @export
 #' @importFrom dplyr %>%
@@ -35,19 +40,19 @@
 #' library(ggplot2)
 #' ggplot(ht %>% filter(!is.na(chlor_a)),
 #' aes(x = lon, y = lat, fill = chlor_a)) + geom_tile()
-hyper_tibble <- function(x, ..., na.rm = TRUE) {
+hyper_tibble <- function(x, ..., na.rm = TRUE, force = FALSE) {
   UseMethod("hyper_tibble")
 }
 #' @name hyper_tibble
 #' @export
-hyper_tibble.character <- function(x, ..., na.rm = TRUE) {
-  tidync(x) %>% hyper_filter(...) %>% hyper_tibble()
+hyper_tibble.character <- function(x, ..., na.rm = TRUE, force = FALSE) {
+  tidync(x) %>% hyper_filter(...) %>% hyper_tibble(na.rm = na.rm, force = force)
 }
 #' @name hyper_tibble
 #' @export
-hyper_tibble.tidync<- function(x, ..., na.rm = TRUE) {
+hyper_tibble.tidync<- function(x, ..., na.rm = TRUE, force = FALSE) {
   
-  slabs <- hyper_array(x, ...)
+  slabs <- hyper_array(x, ..., na.rm = na.rm, force = force)
   if (na.rm) all_na <- Reduce(`&`, lapply(slabs, 
                                           function(a) is.na(as.vector(a))))
   total_prod <- prod(dim(slabs[[1]]))