+++ R/zzz.R
@@ -1,0 +1,13 @@
+tidync_default_options <- function() {
+  list(tidync.large.data.check = TRUE)
+}
+
+
+.onLoad <- function(libname, pkgname) {
+  op <- options()
+  rdo <- tidync_default_options()
+  toset <- !(names(rdo) %in% names(op))
+  if (any(toset)) options(rdo[toset])
+  
+  invisible()
+}